requirements-dev.txt
pytest==4.4.1
pytest-cov==2.6.1
pytest-django==3.4.8
